design: 마이페이지 SaveCourseContainer 크기 조절
