+++ footlog/src/app/globals.css
@@ -294,6 +294,7 @@ button {
 
 .fonts-logLocations {
   color: #333;
+
   font-family: 'Noto Sans KR';
   font-size: 0.875rem;
   font-style: normal;
@@ -305,8 +306,10 @@ button {
   box-shadow: none;
   padding: 0;
 }
+
 .fonts-logOneLocation {
   color: #333;
+
   font-family: 'Noto Sans KR';
   font-size: 1.125rem;
   font-style: normal;
@@ -314,8 +317,10 @@ button {
   line-height: 130%; /* 1.4625rem */
   letter-spacing: -0.03125rem;
 }
+
 .fonts-logContent {
   color: #808080;
+
   font-family: 'Noto Sans KR';
   font-size: 0.875rem;
   font-style: normal;
@@ -323,28 +328,22 @@ button {
   line-height: 1.0625rem; /* 121.429% */
   letter-spacing: -0.03125rem;
 }
-.fonts-mypageNull {
-  color: #808080;
-  font-family: 'Noto Sans KR';
-  font-size: 0.875rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 1.0625rem; /* 121.429% */
-  letter-spacing: -0.03125rem;
-}
+
 .fonts-logDone {
   color: #fff;
+
   text-align: center;
   font-family: 'Noto Sans KR';
   font-size: 1rem;
   font-style: normal;
   font-weight: 700;
   line-height: 130%; /* 1.3rem */
-  line-height: 130%; /* 1.1375rem */
   letter-spacing: -0.03125rem;
 }
+
 .fonts-newLogContent {
   color: #333;
+
   font-family: 'Noto Sans KR';
   font-size: 0.875rem;
   font-style: normal;
